include-linux-timeh-make-nsec_per_sec-64-bit-on-32-bit-architectures-fix

+++ fs/proc/uptime.c
@@ -27,9 +27,9 @@ static int uptime_proc_show(struct seq_file *m, void *v)
 	idle.tv_nsec = rem;
 	seq_printf(m, "%lu.%02lu %lu.%02lu\n",
 			(unsigned long) uptime.tv_sec,
-			(uptime.tv_nsec / (NSEC_PER_SEC / 100)),
+			(uptime.tv_nsec / ((long)NSEC_PER_SEC / 100)),
 			(unsigned long) idle.tv_sec,
-			(idle.tv_nsec / (NSEC_PER_SEC / 100)));
+			(idle.tv_nsec / ((long)NSEC_PER_SEC / 100)));
 	return 0;
 }
